Edited some bugs
